faceswap-engine/requirements.txt
# Private FaceSwap Engine — Python Dependencies
# Install: pip install -r requirements.txt

# Core AI / Face Swap
insightface==0.7.3
onnxruntime-gpu          # Use 'onnxruntime' for CPU-only systems

# Computer Vision
opencv-python>=4.8.0

# Web UI
gradio==4.44.0

# Face Enhancement
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0

# PyTorch (install separately for CUDA support)
# GPU:  pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# CPU:  pip install torch torchvision torchaudio
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Utilities
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
